+++ src/Candoumbe.Pipelines/Components/Workflows/IDoColdfixWorkflow.cs
@@ -2,7 +2,6 @@ using System;
 using System.Threading.Tasks;
 using Nuke.Common;
 using static Nuke.Common.Tools.Git.GitTasks;
-using static Serilog.Log;
 
 namespace Candoumbe.Pipelines.Components.Workflows;
 
@@ -45,7 +44,6 @@ public interface IDoColdfixWorkflow : IDoFeatureWorkflow
             if (!GitRepository.Branch.Like($"{ColdfixBranchPrefix}/*"))
             {
                 AskBranchNameAndSwitchToIt(ColdfixBranchPrefix, ColdfixBranchSourceName);
-                Information($"{EnvironmentInfo.NewLine}Good bye !");
             }
             else
             {