+++ src/Candoumbe.Pipelines/Components/Workflows/IDoFeatureWorkflow.cs
@@ -2,7 +2,6 @@ using System.Threading.Tasks;
 using Nuke.Common;
 using Nuke.Common.Git;
 using static Nuke.Common.Tools.Git.GitTasks;
-using static Serilog.Log;
 
 namespace Candoumbe.Pipelines.Components.Workflows;
 
@@ -48,9 +47,7 @@ public interface IDoFeatureWorkflow : IWorkflow
        {
            if (!GitRepository.IsOnFeatureBranch())
            {
-               Information("Enter the name of the feature. It will be used as the name of the feature/branch (leave empty to exit) :");
                AskBranchNameAndSwitchToIt(FeatureBranchPrefix, sourceBranch: FeatureBranchSourceName);
-               Information("Good bye !");
            }
            else
            {